requirements.txt
sk-video
torch==1.0.1
torchvision==0.2.2